Updates and Some Front End Stuff
Added some comments so the code is a little better explained.

Started with the front end, added a page that will be the login, already getting the values.

After this, i will connect the server to the PostgresSQL database and then using the socket to transfer the login information between client and server.